+++ src/js/main.js
@@ -4,22 +4,8 @@ import spark from '@sparkdesignsystem/spark/es5/spark';
 import sparkPrerender from '@sparkdesignsystem/spark/es5/sparkPrerender';
 import '@sparkdesignsystem/spark/es5/sparkPolyfills';
 
-import { setSpinning, cancelSpinning } from '@sparkdesignsystem/spark';
+require('./buttons');
 
 // initialize Spark
 sparkPrerender();
 spark();
-
-window.startSpinners = (classname) => {
-  const spinningButtons = document.querySelectorAll('.' + classname)
-  spinningButtons.forEach((button) => {
-    setSpinning(button, {});
-  })
-}
-
-window.resetSpinners = (classname) => {
-  const spinningButtons = document.querySelectorAll('.sprk-c-Button--has-spinner.' + classname);
-  spinningButtons.forEach((button) => {
-    cancelSpinning(button);
-  })
-}
