+++ src/pages/buttons.html
@@ -76,7 +76,7 @@
       data-sprk-spinner-variant="dark">
       Quaternary
     </button>
-    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="window.resetSpinners('demo1')">
+    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="resetSpinners('demo1')">
       Stop Spinning
     </button>
     <h2 class="sprk-demo-b-h2">Spinner With Options</h2>
@@ -85,17 +85,17 @@
       data-sprk-spinner-aria-valuetext="Testing valuetext">
       Primary
     </button>
-    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="window.resetSpinners('demo4')">
+    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="resetSpinners('demo4')">
       Stop Spinning
     </button>
     <h2 class="sprk-demo-b-h2">Spinner With Alternative Trigger</h2>
     <button class="sprk-c-Button demo2" aria-live="polite">
       Primary
     </button>
-    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="window.startSpinners('demo2')">
+    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="startSpinners('demo2')">
       Start Spinning
     </button>
-    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="window.resetSpinners('demo2')">
+    <button class="sprk-c-Button sprk-c-Button--tertiary" onclick="resetSpinners('demo2')">
       Stop Spinning
     </button>
   </main>
