--- src/templates/alert/info.hbs
@@ -1,36 +1,0 @@
-<div
- class="sprk-c-Alert sprk-c-Alert--info"
- role="alert"
- data-sprk-alert="container"
- data-id="alert-info-1"
- data-analytics="object.action.event"
->
-  <div class="sprk-c-Alert__content">
-    <svg
-     class="sprk-c-Alert__icon sprk-c-Icon sprk-c-Icon--xl sprk-c-Icon--filled-current-color"
-     viewBox="0 0 64 64"
-     aria-hidden="true"
-    >
-      <use xlink:href="#bell-filled"></use>
-    </svg>
-
-    <p class="sprk-c-Alert__text">
-      This is important information.
-    </p>
-  </div>
-
-  <button
-   class="sprk-c-Alert__icon sprk-c-Alert__icon--dismiss"
-   type="button"
-   title="Dismiss"
-   data-sprk-alert="dismiss"
-  >
-    <svg
-     class="sprk-c-Icon sprk-c-Icon--filled-current-color sprk-c-Icon--stroke-current-color"
-     viewBox="0 0 64 64"
-     aria-hidden="true"
-    >
-      <use xlink:href="#close"></use>
-    </svg>
-  </button>
-</div>
