--- src/templates/alert/no-dismiss.hbs
@@ -1,20 +1,0 @@
-<div
- class="sprk-c-Alert sprk-c-Alert--success"
- role="alert"
- data-sprk-alert="container"
- data-id="alert-success-2"
- data-analytics="object.action.event"
->
-  <div class="sprk-c-Alert__content">
-    <svg
-     class="sprk-c-Alert__icon sprk-c-Icon sprk-c-Icon--xl sprk-c-Icon--filled-current-color"
-     viewBox="0 0 512 512"
-     aria-hidden="true"
-    >
-      <use xlink:href="#check-mark-filled"></use>
-    </svg>
-    <p class="sprk-c-Alert__text">
-      This success alert has no dismiss button.
-    </p>
-  </div>
-</div>
