--- src/templates/alerts.hbs
@@ -1,5 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/alert" text="Alert Docs"}}
-
-{{> './alert/success' }}<br>
-{{> './alert/info' }}<br>
-{{> './alert/fail' }}<br>
