--- src/templates/box.hbs
@@ -1,7 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/box" text="Box Docs"}}
-
-<div class="sprk-o-Box">
-  <p class="sprk-b-TypeBodyTwo">
-    This is a paragraph inside of a box object.
-  </p>
-</div>
