--- src/templates/button/disabled.hbs
@@ -1,7 +1,0 @@
-<button
- class="sprk-c-Button sprk-is-Disabled"
- data-id="button-1"
- disabled
->
-  Button
-</button>
