--- src/templates/button/primary.hbs
@@ -1,6 +1,0 @@
-<button
- class="sprk-c-Button"
- data-id="button-1"
->
-  Button
-</button>
