--- src/templates/button/secondary.hbs
@@ -1,7 +1,0 @@
-<button
- class="sprk-c-Button sprk-c-Button--secondary"
- type="button"
- data-id="button-1"
->
-  Button
-</button>
