--- src/templates/button/tertiary.hbs
@@ -1,6 +1,0 @@
-<button
- class="sprk-c-Button sprk-c-Button--tertiary"
- data-id="button-1"
->
-  Button
-</button>
