--- src/templates/buttons.hbs
@@ -1,8 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/button" text="Button Docs"}}
-
-<p>
-{{> 'button/primary' }} &nbsp;
-{{> 'button/secondary' }} &nbsp;
-{{> 'button/tertiary' }} &nbsp;
-{{> 'button/disabled' }} &nbsp;
-</p>
