--- src/templates/card/base.hbs
@@ -1,6 +1,0 @@
-<div class="sprk-c-Card sprk-o-Stack">
-
-  <div class="sprk-o-Stack__item sprk-c-Card__content sprk-o-Stack sprk-o-Stack--large">
-
-  </div>
-</div>
