--- src/templates/card/highlighted-header.hbs
@@ -1,22 +1,0 @@
-<div
- class="sprk-c-Card sprk-c-Card--standout sprk-o-Stack"
- data-id="card-12"
->
-  <div class="sprk-o-Stack__item sprk-c-Card__header sprk-o-Stack sprk-o-Stack--medium">
-    <h3 class="sprk-b-TypeDisplaySeven sprk-o-Stack__item sprk-u-Color--white">
-      Description
-    </h3>
-
-    <h4 class="sprk-b-TypeDisplayFive sprk-o-Stack__item sprk-u-Color--white">
-      Card Title
-    </h4>
-  </div>
-
-  <div class="sprk-o-Stack__item sprk-o-Stack sprk-o-Stack--medium sprk-c-Card__content">
-    <p class="sprk-b-TypeBodyTwo sprk-o-Stack__item">
-      Lorem ipsum dolor sit amet, doctus invenire vix te. Facilisi
-      perpetua an pri, errem commune mea at, mei prima tantas
-      signiferumque at. Numquam.
-    </p>
-  </div>
-</div>
