--- src/templates/card/standout.hbs
@@ -1,7 +1,0 @@
-<div
- class="sprk-c-Card sprk-c-Card--standout sprk-o-Stack"
- data-id="card-12"
->
-  <div class="sprk-o-Stack__item sprk-c-Card__content sprk-o-Stack sprk-o-Stack--medium">
-  </div>
-</div>
