--- src/templates/card/teaser-alt.hbs
@@ -1,28 +1,0 @@
-<div
- class="sprk-c-Card sprk-o-Stack"
- data-id="card-2"
->
-  <h3 class="sprk-o-Stack__item sprk-c-Card__content sprk-b-TypeDisplayFive">
-    Card Teaser Extended Title
-  </h3>
-
-  <a href="#" class="sprk-o-Stack__item">
-    <img
-      alt="Spark placeholder image"
-      src="./src/assets/placeholder.jpg">
-  </a>
-
-  <div class="sprk-o-Stack__item sprk-c-Card__content sprk-o-Stack sprk-o-Stack--large">
-    <p class="sprk-b-TypeBodyTwo sprk-o-Stack__item">
-      Lorem ipsum dolor sit amet, doctus invenire
-      vix te. Facilisi perpetua an pri, errem commune
-      mea at, mei prima tantas signiferumque at. Numquam.
-    </p>
-
-    <div class="sprk-o-Stack__item">
-      <a href="#" class="sprk-c-Button">
-        Learn More
-      </a>
-    </div>
-  </div>
-</div>
