--- src/templates/card/teaser.hbs
@@ -1,29 +1,0 @@
-<div
- class="sprk-c-Card sprk-o-Stack"
- data-id="card-1"
->
-  <a href="#" class="sprk-o-Stack__item">
-    <img
-      class="sprk-c-Card__media"
-      alt="Spark placeholder image."
-      src="./src/assets/placeholder.jpg">
-  </a>
-
-  <div class="sprk-o-Stack__item sprk-c-Card__content sprk-o-Stack sprk-o-Stack--large">
-
-    <h3 class="sprk-b-TypeDisplayFive sprk-o-Stack__item">
-      Title
-    </h3>
-
-    <p class="sprk-b-TypeBodyTwo sprk-o-Stack__item">
-      Lorem ipsum dolor sit amet, doctus invenire vix te. Facilisi perpetua an pri, errem commune mea at, mei prima tantas signiferumque at. Numquam.
-    </p>
-
-    <div class="sprk-o-Stack__item">
-      <a href="#" class="sprk-c-Button">
-        Learn More
-      </a>
-    </div>
-
-  </div>
-</div>
