--- src/templates/centered-column.hbs
@@ -1,7 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/centered-column" text="Centered Column Docs"}}
-
-<div class="sprk-o-CenteredColumn">
-  <p class="sprk-b-TypeBodyTwo">
-    This is a paragraph inside of a centered column.
-  </p>
-</div>
