--- src/templates/divider.hbs
@@ -1,3 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/divider" text="Divider Docs"}}
-
-<span class="sprk-c-Divider" data-id="divider-1"></span>
