--- src/templates/flag.hbs
@@ -1,37 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/flag" text="Flag Docs"}}
-
-<div class="sprk-o-Flag">
-  <div class="sprk-o-Flag__figure">
-    <img alt="Provide useful alternative text" src="https://spark-assets.netlify.app/spark-logo-mark.svg" />
-  </div>
-  <div class="sprk-o-Flag__body">
-    <p>
-      Lorem ipsum dolor. Sit amet pede. Tempus donec et.
-      Suspendisse id inventore integer eum non enim diam habitant.
-    </p>
-  </div>
-</div>
-
-<div class="sprk-o-Flag sprk-o-Flag--rev">
-  <div class="sprk-o-Flag__figure">
-    <img alt="Provide useful alternative text" src="https://spark-assets.netlify.app/spark-logo-mark.svg" />
-  </div>
-  <div class="sprk-o-Flag__body">
-    <p>
-      Lorem ipsum dolor. Sit amet pede. Tempus donec et.
-      Suspendisse id inventore integer eum non enim diam habitant.
-    </p>
-  </div>
-</div>
-
-  <div class="sprk-o-Flag sprk-o-Flag--stacked">
-    <div class="sprk-o-Flag__body">
-      <p>
-        Lorem ipsum dolor. Sit amet pede. Tempus donec et.
-        Suspendisse id inventore integer eum non enim diam habitant.
-      </p>
-    </div>
-    <div class="sprk-o-Flag__figure">
-      <img alt="Provide useful alternative text" src="https://spark-assets.netlify.app/spark-logo-mark.svg" />
-    </div>
-  </div>
