--- src/templates/icons.hbs
@@ -1,5 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/icon" text="Icon Docs"}}
-
-<svg class="sprk-c-Icon  sprk-c-Icon--xl" viewBox="0 0 100 100">
-  <use xlink:href="#access" />
-</svg>
