--- src/templates/link/base.hbs
@@ -1,1 +1,0 @@
-<a class="sprk-b-Link" href="#nogo" data-id="link-1" data-analytics="object.action.event">Base Link</a>
