--- src/templates/link/disabled.hbs
@@ -1,1 +1,0 @@
-<a class="sprk-b-Link sprk-b-Link--disabled" href="#nogo" data-id="link-3" data-analytics="object.action.event">Disabled Link</a>
