--- src/templates/link/icon.hbs
@@ -1,11 +1,0 @@
-<a
-  class="sprk-b-Link sprk-b-Link--has-icon sprk-b-Link--simple"
-  href="#nogo"
-  data-id="link-1"
-  data-analytics="object.action.event">
-  <svg class="sprk-c-Icon sprk-c-Icon--l sprk-u-mrs sprk-c-Icon--filled-current-color" viewBox="0 0 100 100">
-    <use xlink:href="#communication" />
-  </svg>
-
-  Icon Link
-</a>
