--- src/templates/link/simple.hbs
@@ -1,1 +1,0 @@
-<a class="sprk-b-Link sprk-b-Link--simple" href="#nogo" data-id="link-2" data-analytics="object.action.event">Simple Link</a>
