--- src/templates/links.hbs
@@ -1,8 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/link" text="Link Docs"}}
-
-<p>
-{{> 'link/base' }} &nbsp;
-{{> 'link/simple' }} &nbsp;
-{{> 'link/icon' }} &nbsp;
-{{> 'link/disabled' }} &nbsp;
-</p>
