--- src/templates/masthead.hbs
@@ -1,5 +1,0 @@
-{{#if (eq variant 'extended') }}
-  {{> 'masthead/extended.hbs' }}
-{{else }}
-  {{> 'masthead/base.hbs' }}
-{{/if }}
