--- src/templates/modal/info.hbs
@@ -1,44 +1,0 @@
-<div
- class="sprk-c-Modal sprk-u-Display--none"
- role="dialog"
- tabindex="-1"
- aria-labelledby="modalInfoHeading"
- aria-modal="true"
- aria-describedby="modalInfoContent"
- data-sprk-modal="exampleInfoModal"
- data-id="modal-info-1"
->
-  <div class="sprk-o-Stack sprk-o-Stack--large">
-    <header class="sprk-o-Stack__item sprk-c-Modal__header">
-      <h2 class="sprk-c-Modal__heading sprk-b-TypeDisplayFour" id="modalInfoHeading">
-        Info Modal
-      </h2>
-
-      <button
-       class="sprk-c-Modal__icon"
-       data-sprk-modal-cancel="exampleInfoModal"
-       type="button"
-       aria-label="Close Modal"
-      >
-        <svg
-         class="sprk-c-Icon"
-         viewBox="0 0 64 64"
-         aria-hidden="true"
-         focusable="false"
-        >
-          <use xlink:href="#close"></use>
-        </svg>
-      </button>
-    </header>
-
-    <div class="sprk-o-Stack__item sprk-c-Modal__body">
-      <p
-       class="sprk-b-TypeBodyTwo"
-       id="modalInfoContent"
-      >
-        This is some content that is in a Modal. It explains what the Modal is for.
-        There will also be a way to close the modal.
-      </p>
-    </div>
-  </div>
-</div>
