--- src/templates/modal/wait.hbs
@@ -1,35 +1,0 @@
-<div
- class="sprk-c-Modal sprk-c-Modal--wait sprk-u-Display--none"
- role="dialog"
- tabindex="-1"
- aria-labelledby="modalWaitHeading"
- aria-modal="true"
- aria-describedby="modalWaitContent"
- data-sprk-modal="exampleWaitModal"
- data-sprk-modal-type="wait"
- data-id="modal-wait-1"
->
-  <div class="sprk-o-Stack sprk-o-Stack--large">
-    <header class="sprk-o-Stack__item sprk-c-Modal__header">
-      <h2 class="sprk-c-Modal__heading sprk-b-TypeDisplayFour" id="modalWaitHeading">
-        Please Wait...
-      </h2>
-    </header>
-
-    <div class="sprk-o-Stack__item sprk-c-Modal__body sprk-o-Stack sprk-o-Stack--medium">
-      <div class="sprk-o-Stack__item sprk-c-Spinner sprk-c-Spinner--circle sprk-c-Spinner--large sprk-c-Spinner--primary"></div>
-      <p
-       class="sprk-o-Stack__item sprk-b-TypeBodyTwo"
-       id="modalWaitContent"
-      >
-        This modal will close shortly for demonstration purposes.
-      </p>
-    </div>
-  </div>
-</div>
-
-<div
- data-sprk-modal-mask="true"
- class="sprk-c-ModalMask sprk-u-Display--none"
- tabindex="-1"
-></div>
