--- src/templates/modals.hbs
@@ -1,19 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/patterns/components/modals" text="Modal Docs"}}
-
-<button class="sprk-c-Button" data-sprk-modal-trigger="exampleChoiceModal" type="button">
-  Open Choice Modal
-</button>
-
-<button class="sprk-c-Button" data-sprk-modal-trigger="exampleInfoModal" type="button">
-  Open Info Modal
-</button>
-
-<button class="sprk-c-Button" data-sprk-modal-trigger="exampleWaitModal" type="button">
-  Open Wait Modal
-</button>
-
-{{> 'modal/info' }}
-{{> 'modal/choice' }}
-{{> 'modal/wait' }}
-
-<div data-sprk-modal-mask="true" class="sprk-c-ModalMask sprk-u-Display--none" tabindex="-1"></div>
