--- src/templates/pagination.hbs
@@ -1,66 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/modal" text="Pagination Docs"}}
-
-<nav aria-label="Pagination Navigation" data-sprk-pagination="default" data-id="pagination-1">
-  <ul class="sprk-c-Pagination sprk-o-HorizontalList sprk-o-HorizontalList--spacing-medium">
-    <li>
-      <a class="sprk-c-Pagination__icon sprk-b-Link sprk-b-Link--disabled sprk-b-Link--plain" href="#nogo"
-        data-sprk-pagination="prev">
-        <span class="sprk-u-ScreenReaderText">Prev</span>
-        <svg class="sprk-c-Icon sprk-c-Icon--filled-current-color sprk-c-Icon--stroke-current-color" viewBox="0 0 64 64">
-          <use xlink:href="#chevron-left"></use>
-        </svg>
-      </a>
-    </li>
-
-    <li>
-      <a class="sprk-c-Pagination__link sprk-c-Pagination__link--current sprk-b-Link" href="#nogo"
-        data-sprk-pagination="item" aria-current="true">
-        1
-      </a>
-    </li>
-
-    <li>
-      <a class="sprk-c-Pagination__link sprk-b-Link" href="#nogo" data-sprk-pagination="item">
-        2
-      </a>
-    </li>
-
-    <li>
-      <a class="sprk-c-Pagination__link sprk-b-Link" href="#nogo" data-sprk-pagination="item">
-        3
-      </a>
-    </li>
-
-    <li>
-      <a class="sprk-c-Pagination__icon sprk-b-Link sprk-b-Link--plain" href="#nogo" data-sprk-pagination="next">
-        <span class="sprk-u-ScreenReaderText">Next</span>
-        <svg class="sprk-c-Icon sprk-c-Icon--filled-current-color sprk-c-Icon--stroke-current-color" viewBox="0 0 64 64">
-          <use xlink:href="#chevron-right"></use>
-        </svg>
-      </a>
-    </li>
-  </ul>
-</nav>
-
-<nav aria-label="Pagination Navigation" data-id="pagination-3">
-  <ul class="sprk-c-Pagination sprk-o-HorizontalList sprk-o-HorizontalList--spacing-large">
-    <li>
-      <a class="sprk-c-Pagination__icon sprk-b-Link sprk-b-Link--plain" href="#nogo">
-        <svg class="sprk-c-Icon sprk-c-Icon--filled-current-color sprk-c-Icon--stroke-current-color" viewBox="0 0 64 64">
-          <use xlink:href="#chevron-left"></use>
-        </svg>
-        <span class="sprk-u-ScreenReaderText">Prev</span>
-      </a>
-    </li>
-
-    <li>
-      <a class="sprk-c-Pagination__icon sprk-b-Link sprk-b-Link--plain" href="#nogo">
-        <svg class="sprk-c-Icon sprk-c-Icon--filled-current-color sprk-c-Icon--stroke-current-color" viewBox="0 0 64 64">
-          <use xlink:href="#chevron-right"></use>
-        </svg>
-        <span class="sprk-u-ScreenReaderText">Next</span>
-      </a>
-    </li>
-  </ul>
-</nav>
-
