--- src/templates/stack.hbs
@@ -1,10 +1,0 @@
-{{> './_doclink' href="https://sparkdesignsystem.com/using-spark/components/stack" text="Stack Docs"}}
-
-<div class="sprk-o-Stack">
-  <div class="sprk-o-Stack__item">
-    Item 1
-  </div>
-  <div class="sprk-o-Stack__item">
-    Item 2
-  </div>
-</div>
