--- src/templates/typography/body-one.hbs
@@ -1,17 +1,0 @@
-<p class="sprk-b-TypeBodyOne sprk-u-mbl">Body One - Lorem ipsum dolor sit amet, consectetur
-  adipiscing elit. Sed gravida urna quis nulla ultrices, sed efficitur risus elementum. Vivamus
-  semper ex a lorem sodales, rhoncus rhoncus augue cursus. Sed vestibulum felis vel euismod
-  aliquet. Suspendisse sit amet euismod nibh, eget fermentum arcu. Aliquam lacinia, sem eu
-  ultricies auctor, velit nunc commodo sapien, pharetra maximus augue dui eget dui.
-
-  <span data-sprk-tooltip="container" class="sprk-c-Tooltip__container">
-    <button data-sprk-tooltip="trigger" class="sprk-c-Tooltip__trigger" aria-labelledby="tooltip_1">
-      <svg class="sprk-c-Icon sprk-c-Icon--filled" aria-hidden="true">
-        <use xlink:href="#information-filled" />
-      </svg>
-    </button>
-    <span data-sprk-tooltip="content" class="sprk-c-Tooltip" aria-hidden="true" id="tooltip_1" role="tooltip">
-      Tooltips provide a short text description of information about a feature.
-    </span>
-  </span>
-</p>
