--- src/templates/typography/h1-title.hbs
@@ -1,1 +1,0 @@
-<h1 class="sprk-b-TypeDisplayOne sprk-b-PageTitle sprk-u-mbl">Display One Title!</h1>
