--- src/templates/typography/h5.hbs
@@ -1,1 +1,0 @@
-<h5 class="sprk-b-TypeDisplaySix sprk-u-mbl">Heading 5 - Lorem ipsum dolor amet ennui lomo health goth retro synth twee skateboard pinterest locavore seitan man bun. Mixtape franzen yr, affogato marfa vegan af lyft tofu selfies semiotics. </h5>
